A4/db/requirements.txt
pandas
psycopg2-binary==2.9.1
fastparquet